rename api_keys_manager
