+++ services/director-v2/src/simcore_service_director_v2/modules/osparc_variables/__init__.py
@@ -1,4 +1,4 @@
-from ._api_keys_manager import safe_remove_api_key_and_secret
+from .api_keys_manager import safe_remove_api_key_and_secret
 
 assert safe_remove_api_key_and_secret  # nosec
