+++ services/director-v2/src/simcore_service_director_v2/modules/osparc_variables/api_keys_manager.py
